refactor: add fixtures
