+++ __tests__/routers/groupRouter/fixtures/group.ts
@@ -1,0 +1,12 @@
+import { v4 as uuidv4 } from 'uuid';
+import { Group } from '../../../../src/models';
+
+const id = uuidv4();
+const group: Group = {
+  id: id,
+  name: 'test group name',
+  permissions: ['READ', 'WRITE', 'DELETE', 'SHARE', 'UPLOAD_FILES'],
+  users: [],
+};
+
+export { id, group };
